server/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
PyJWT==2.6.0
pythainlp==3.1.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.4.0
